IMPALA-8820: fix start-impala-cluster catalogd startup
The catalogd process sometimes changes its name to "main"
after an ubuntu 16.04 update.

This avoids the issue by checking the first element of the
command line instead, which should reflect the binary
that was executed more reliably.

Testing:
This failed consistently before the change and now passes consistently
on my development machine.

Change-Id: Ib9396669481e4194beb6247c8d8b6064cb5119bb
Reviewed-on: http://gerrit.cloudera.org:8080/13971
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>