IMPALA-8781: Fix TestResultSpooling::test_multi_batches
Prefix the query in TestResultSpooling::test_multi_batches with the
database name. This was causing the Dockerized tests to fail. I double
checked what other tests do and all the ones I saw either switch to the
appropriate database or prefix the table name using the database name.
The latter seemed more straightforward.

I was not able to re-produce this locally, and its odd that this only
affected the Dockerized tests (even more odd is that it seems to either
be intermittent, or only affecting Dockerized tests triggered by
gerrit-verify-dryrun-external). Regardless, it is a straightforward fix
that makes the TestResultSpooling::test_multi_batches consistent with
the rest of the tests.

Testing:
* Ran test_result_spooling.py locally using both bin/impala-py.test
and tests/run-tests.py.

Change-Id: I939eedba37003f5c720cea96e5c3532e2cc6312c
Reviewed-on: http://gerrit.cloudera.org:8080/14022
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>