IMPALA-8600: Fix test_acid_compaction
Apparently Hive doesn't update the writeIds on a transactional table
after compaction. This breaks an assumption made in the REFRESH table
logic that only does an actual refresh when the HMS writeId is
different than the one cached locally. As a result the non-partitioned
ACID tables aren't refreshed when a REFRESH table is invoked in Impala
right after a major compaction.

Change-Id: I58b79f8864b31e18eca818032ad5a9af954913f6
Reviewed-on: http://gerrit.cloudera.org:8080/14027
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>