IMPALA-8848: fix UNION missing input cardinality bug
If a UNION has children, and none of those children has
a known cardinality, then we can make no reasonable
estimate of the output cardinality, so the planner
should consider the output cardinality to be unknown.

The previous behaviour was to report a cardinality of
0, which is unsafe because the planner may make further
decisions under the incorrect assumption that the output
of the UNION is tiny.

Testing:
An existing CardinalityTest already tested this but had
the wrong estimate.

Change-Id: Ic3ed670ffb685d8ff24824933ca303f3219737bb
Reviewed-on: http://gerrit.cloudera.org:8080/14036
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>