IMPALA-8124: Modify TestWebPage::test_catalog to avoid flakiness.
This test scrapes the /catalog webpage for metrics. This can
occasionally run into a race condition when a table lock is
being held for functional*.* tables and the test tries to get
the metrics. This result in a failed assert for the metrics.
This change rewrites the test to create new tables with
unique_database fixture to avoid flakiness.

Also, seperated out an assert with AND into two different asserts
so it's easy to track which one failed.

Testing:
Ran multiple runs of this test locally while refreshing the table.

Change-Id: I341bf25baf8d9316a21a9eff860de84b33afd12f
Reviewed-on: http://gerrit.cloudera.org:8080/14075
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>