IMPALA-8850: is_percent not initialized in TmpFileMgr parser
This bug was only detected on ASAN. The is_percent variable
was not initialised and is not set in ParseMemSpec() when it
does an early exit for an empty string. This caused
TmpFileMgr to consider the empty memory spec after the colon
to be invalid.

Testing:
This reliably reproduced in TmpFileMgrTest under ASAN.
Initialising is_percent fixed the issue.

Added more detailed tests for ParseMemSpec() that initialise
is_percent to avoid similar bugs.

Change-Id: I4465944c7043c1fc6b33db9693ec290c24e3ed19
Reviewed-on: http://gerrit.cloudera.org:8080/14040
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>