IMPALA-8867: Deflake test_auto_scaling, improve logging
test_auto_scaling sometimes failed to reach the required query rate
within the expected time. This change increases the timeout that we wait
for the query rate to increase. It also adds logging for the maximum
query rate observed to help with debugging in case the issue still
occurs.

Testing: I looped this for a day on my desktop box and did not observe
any issues.

Change-Id: I22c43618a40ff197784add69223359e23fa1bdec
Reviewed-on: http://gerrit.cloudera.org:8080/14074
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>