IMPALA-8868: Fix 401 response when LDAP and Kerberos are enabled
When both kerberos and ldap auth are enabled and an http request is
not successfully authenticated, THttpServer only sends the
'WWW-Authenticate: Basic' challenge and doesn't send the
'WWW-Authenticate: Negotiate' challenge, which can cause clients that
want to connect with kerberos to fail to authenticate.

This patch fixes this to send both challenges.

Testing:
- Manually tested in a cluster with both Kerberos and LDAP enabled on
  Impala with connections proxied through Apache Knox, which would
  previously fail.

Change-Id: I138f33783bfd0f8f9b8db242589a9cc75cfd392a
Reviewed-on: http://gerrit.cloudera.org:8080/14077
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>