IMPALA-8895: Expose daemon health endpoint
This change exposes the daemon health of an impalad via an HTTP endpoint
'/healthz'. If the server is healthy, this endpoint will return HTTP
code 200 (OK). If it is unhealthy, it will return 503 (Service
Unavailable). With this change, the endpoint is equivalent to the
'impala-server.ready' metric.

Testing: I added a test in test_web_pages.py and observed manually that
the endpoint returns 503 while starting up.

Change-Id: I05889e21bbb1b97b23fad9928b9e61e5dbff2615
Reviewed-on: http://gerrit.cloudera.org:8080/14146
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>