IMPALA-8901: Fix # links on /catalog page.
After IMPALA-7935, the database links on the top of /catalog page
that use # links to jump to the part of the page corresponding to
the particular database is broken because the id is ommitted. This is
fixed by moving id to a tag which doesn't change if local catalog is
turned on.

Testing:
Manually tested the links are working in both local catalog mode and
V1 mode.

Change-Id: I895742dcaeaf71ecf097b0fec69cc19a3b6f86f4
Reviewed-on: http://gerrit.cloudera.org:8080/14153
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>