IMPALA-8907: TestResultSpooling.test_slow_query is flaky
De-flake test_result_spooling.py::TestResultsSpooling::test_slow_query
by increasing the delay in RowBatch production. This patch makes two
fixes to ensure that RowBatchGetWaitTime is a non-zero value:

* Add the DELAY DEBUG_ACTION to the SCAN_NODE rather than the
EXCHANGE_NODE. Since the EXCHANGE_NODE only processes a few rows, adding
the delay to the SCAN_NODE decreases the rate at which results are
produced.

* Wait for all rows to be fetched before checking if RowBatchGetWaitTime
is in the profile. This fixes a possible race condition where the fetch
thread was not able to issue the fetch RPC before the test checked if
RowBatchGetWaitTime was in the runtime profile.

Testing:
* Looped test_slow_query for several hours with 0 failures

Change-Id: Idcb4a6b38f85a9497f80e2674e1c1fa512be5940
Reviewed-on: http://gerrit.cloudera.org:8080/14170
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>