IMPALA-8897 (part 2): Fix javascript to work with Knox integration
This patch qualifies all urls accessed via javascript in the webui
with the appropriate host:port in order to allow these urls to work
when the connection to the webui is being proxied through Apache Knox.

It accomplishes this with the function make_url(), which takes the
href of a link from the page, which may have been rewritten by Knox,
and appends a path to it.

This patch also fixes on issue on the /admissions page, where
resetting a pool's stats could fail due to the page being reloaded
before the reset is executed. Fixed by moving the call to reload to
the completion callback for the 'reset' endpoint.

Testing:
- Added a regex to test_knox_compatability that performs a rough check
  for places in the tmpl files where urls are used in javascript
  without calling make_url().

Change-Id: I3de9fd1bbb8bb38ce63b3160fcafd33eb0530581
Reviewed-on: http://gerrit.cloudera.org:8080/14173
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>