IMPALA-8912: Avoid sampling hbase table twice for HBaseScanNode
HBaseScanNode#computeStats do sampling on the target hbase table to
estimate cardinality and row size. The sampling can be time consuming
since it requires HBase RPCs. As mentioned in the JIRA description,
HBaseScanNode#computeStats can be called twice if it's the root of the
single node plan. The sampling can be skipped in the second pass since
startKey_, endKey_ and keyConjucts_ are not changed.

inputCardinality_ will have a valid value if the sampling has been done
without errors. We check this to determine whether we can skip sampling
in the second pass.

Tests:
 - Testing in my local env with TRACE log level. Observing the sampling
be done only once.
 - Run CORE tests.

Change-Id: I4650483c0504128048630714e993b481737fd1e2
Reviewed-on: http://gerrit.cloudera.org:8080/14167
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>