IMPALA-8915: reapply IMPALA-8901 fix
This reapplies the below fix and resolves conflicts correctly.

IMPALA-8901: Fix # links on /catalog page.

After IMPALA-7935, the database links on the top of /catalog page
that use # links to jump to the part of the page corresponding to
the particular database is broken because the id is ommitted. This is
fixed by moving id to a tag which doesn't change if local catalog is
turned on.

Testing:
Manually tested the links are working in both local catalog mode and
V1 mode.

Change-Id: I42ea1ec4c386f2eb45f370cd35a104b49786a48c
Reviewed-on: http://gerrit.cloudera.org:8080/14176
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>