IMPALA-8921: Use short name for Ranger grant/revoke requests
For certain grant/revoke Ranger commmands, we ended up passing the
full name which is a problem when kerberos is enabled. Ranger expects
the short name during authorization.

Testing: We do not have test coverage with kerberos enabled, so I
inspected the code manually to make sure we are using getShortName()
everywhere.

Change-Id: I3dc1bf55d50dc2e5fa6e07f16644f0a2773f2d23
Reviewed-on: http://gerrit.cloudera.org:8080/14185
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>