IMPALA-3926: part 1: bump toolchain for rpath fixes
This bumps the toolchain to a version that sets up relative
rpaths correctly, removing the need to set LD_LIBRARY_PATH
when calling binaries or using shared objects from the toolchain.

rpaths embedded in the binaries and shared objects take precedence
over LD_LIBRARY_PATH, so the new binaries will link against the
shared libraries from the toolchain, as expected.

This does not attempt to clean up LD_LIBRARY_PATH yet -
that would not work for developers who have older versions
of the binaries on their system. We will do that later,
but give devs a window to switch over to the new toolchain.

Change-Id: I2abe513ebf4f459aa3e0c65c77ce7be8b8ca1221
Reviewed-on: http://gerrit.cloudera.org:8080/14274
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>