IMPALA-8647: fix round-to-zero in planner estimates
Zero estimates in query plans can be quite dangerous
(for causing bad plans) because they propagate up the
plan tree, causing many more estimates to be zero.

E.g. if the estimate for the right side of the join
becomes 0, but the left side of the join has 1 trillion
rows, the planner will think that the join produces
zero rows and make further decisions accordingly.
This could be catastrophic if the right side of the
join produces even a single row.

This adds some helper functions that avoid rounding
to zero when multiplying selectivities and uses them
in places where selectivities were computed. The scan
nodes already bumped up the estimates anyway, so this
doesn't change cardinality estimates or plans, but
SelectNode and AggregationNode will not round to zero
after this change.

I looked at using Math.ceil() instead of Math.round(),
but it changed estimates slightly in a lot of plans,
which didn't seem worth the hassle.

Testing:
Added targeted cardinality tests for SelectNode and
AggregationNode.

Updated and sanity checked planner tests where estimates
changed.

Change-Id: I148e9f1aede0a1e99b875b0e6af534f4bccf49b7
Reviewed-on: http://gerrit.cloudera.org:8080/14289
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>