IMPALA-2312: simplify stopwatch ElapsedTime()
The stopwatches had two subtly different ElapsedTime()
interfaces, one that included previous running time
and one that didn't. They do not appear to be both
needed. I looked through all the uses and in all
cases using the old TotalElapsedTime() in place of
ElapsedTime() is equivalent, either because the
stopwatch is stopped when ElapsedTime() is called
or because Stop() is never called.

Change-Id: I004833681248de65ce709fc746269686507aba54
Reviewed-on: http://gerrit.cloudera.org:8080/14293
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>