IMPALA-8899, IMPALA-8898: Add cookie support to the webui
This patches takes the machinery for generating and checking cookies
for authentication that was added in IMPALA-8584 and applies to the
webui.

It also fixes an issue where some clients, for example Knox, may
return the cookie value surrounded by "".

It adds metrics for both SPNEGO auth success/failure and cookie auth
success/failure to the webserver.

This patch also fixes IMPALA-8898 by returning an empty cookie with a
Max-Age of 0 on requests where an invalid cookie was provided to
indicate to the client that the cookie should be deleted.

Testing:
- Added a test that uses curl to access the webserver with SPNEGO
  enabled while storing and using cookies. This test only runs when
  curl is present and has the necessary options enabled, which is
  generally not the case in our automated testing runs.

Change-Id: I30788e0539627ee6154ad8183b124947c5da8ef4
Reviewed-on: http://gerrit.cloudera.org:8080/14339
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>